커뮤니티 <2024-1-27> 1. 좋아요 리스트에 출력 기능 구현
