--- src/main/java/com/multi/roadpet/lounge/LoungeDAOInterface.java
@@ -1,25 +1,0 @@
-package com.multi.roadpet.lounge;
-
-import java.util.List;
-
-public interface LoungeDAOInterface {
-
-	void insert(LoungeVO loungeVO);
-
-	void update(LoungeVO loungeVO);
-
-	void delete(LoungeVO loungeVO);
-
-	List<LoungeVO> list(LoungePageVO loungePageVO);
-
-	LoungeVO one(LoungeVO loungeVO);
-
-	int countAll();
-
-	int countSearch(String keyWord, String searchType);
-	
-	void replyCount(int lounge_id, int rpCntUpDown);
-	
-	List<LoungeVO> bestList(LoungeVO loungeVO);
-
-}
