+++ src/main/java/com/multi/roadpet/lounge/LoungeReplyDAO.java
@@ -2,40 +2,14 @@ package com.multi.roadpet.lounge;
 
 import java.util.List;
 
-import org.mybatis.spring.SqlSessionTemplate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
+public interface LoungeReplyDAO {
 
-@Repository
-public class LoungeReplyDAO implements LoungeReplyDAOInterface {
-		
-		@Autowired
-		SqlSessionTemplate my;
-		
-		@Override
-		public int rpInsert(LoungeReplyVO lngRpVO) {
-			 return my.insert("loungeReply.insert", lngRpVO);
-		}
-		
-		@Override
-		public List<LoungeReplyVO> list(int reply_id) {
-			return my.selectList("loungeReply.list", reply_id);
-		
-		}
-		
-		@Override
-		public int rpUpdate(LoungeReplyVO lngRpVO) {
-			return my.update("loungeReply.update", lngRpVO);
-		}
-		
-		@Override
-		public int rpDelete(LoungeReplyVO lngRpVO) {
-			return my.delete("loungeReply.delete", lngRpVO);
-		}
-		
-		@Override
-		public void rpOne(LoungeReplyVO lngRpVO) {
-			my.selectOne("loungeReply.one", lngRpVO);
-		}
-		
+	int rpInsert(LoungeReplyVO lngRpVO);
+
+	List<LoungeReplyVO> list(int reply_id);
+
+	int rpUpdate(LoungeReplyVO lngRpVO);
+
+	int rpDelete(LoungeReplyVO lngRpVO);
+
 }
