--- src/main/java/com/multi/roadpet/lounge/LoungeReplyServiceInterface.java
@@ -1,17 +1,0 @@
-package com.multi.roadpet.lounge;
-
-import java.util.List;
-
-public interface LoungeReplyServiceInterface {
-
-	int rpInsert(LoungeReplyVO lngRpVO);
-
-	List<LoungeReplyVO> list(int reply_id);
-
-	int rpUpdate(LoungeReplyVO lngRpVO);
-
-	int rpDelete(LoungeReplyVO lngRpVO);
-
-	void rpOne(LoungeReplyVO lngRpVO);
-
-}
