+++ src/main/java/com/multi/roadpet/lounge/LoungeService.java
@@ -1,54 +1,22 @@
 package com.multi.roadpet.lounge;
 
+import java.util.HashMap;
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+public interface LoungeService {
 
+	void insert(LoungeVO loungeVO);
 
-@Service
-public class LoungeService implements LoungeServiceInterface {
+	void update(LoungeVO loungeVO);
+
+	void delete(LoungeVO loungeVO);
+
+	List<LoungeVO> list(HashMap<String, Object> loungeMap);
+
+	LoungeVO one(LoungeVO loungeVO);
+
+	int pageCount(String keyWord, String searchType);
 	
-	@Autowired
-	LoungeDAOInterface loungeDAO;
-	
-	@Override
-	public void insert(LoungeVO loungeVO) {
-		loungeDAO.insert(loungeVO);
-	}
-	
-	@Override
-	public void update(LoungeVO loungeVO) {
-		loungeDAO.update(loungeVO);
-	}
-	
-	@Override
-	public void delete(LoungeVO loungeVO) {
-		loungeDAO.delete(loungeVO);
-	}
-	
-	@Override
-	public List<LoungeVO> list(LoungePageVO loungePageVO) {	
-		
-		return loungeDAO.list(loungePageVO);
-	}
-	
-	@Override
-	public LoungeVO one(LoungeVO loungeVO) {
-		return loungeDAO.one(loungeVO);
-	}
-	
-	@Override
-	public int pageCount(String keyWord, String searchType) {
-		if (keyWord == null || keyWord.trim().equals("") && searchType == null) {
-		 return loungeDAO.countAll();
-		}else {
-		 return loungeDAO.countSearch(keyWord, searchType);
-		}
-	}
-	
-	@Override
-	public List<LoungeVO> bestList(LoungeVO loungeVO) {
-		return loungeDAO.bestList(loungeVO);
-	}
+	List<LoungeVO> bestList(LoungeVO loungeVO);
+
 }
