+++ src/main/java/com/multi/roadpet/story/PetStoryVO.java
@@ -11,12 +11,7 @@ public class PetStoryVO {
 	private int story_private;
 	
 
-	public int getUser_id() {
-		return user_id;
-	}
-	public void setUser_id(int user_id) {
-		this.user_id = user_id;
-	}
+
 	public int getStory_no() {
 		return story_no;
 	}
