+++ src/main/resources/mapper/loungeReplyMapper.xml
@@ -2,36 +2,39 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 	<mapper namespace="loungeReply">
+	
+	<!-- 댓글 INSERT Query -->
 	<insert id="insert" parameterType="loungeReplyVO" useGeneratedKeys="true" keyProperty="reply_id">
-		insert into lounge_reply 
-		values (
-			null,
-			#{reply_oriid},
-			#{user_id},
-			#{reply_content},
-			now()
-			)
+	    INSERT INTO lounge_reply 
+	    VALUES (
+	        null,
+	        #{reply_oriid},
+	        #{user_id},
+	        #{reply_content},
+	        NOW(),
+	        #{reply_writer}
+	    )
 	</insert>
 	
+	<!-- 댓글 list를 가져오는 SELECT Query -->
 	<select id="list" parameterType="int" resultType="loungeReplyVO">
-		select * from lounge_reply
-		where reply_oriid = #{reply_id}
-		order by reply_id desc
+	    SELECT * FROM lounge_reply
+	    WHERE reply_oriid = #{reply_id}
+	    ORDER BY reply_id DESC
 	</select>
 	
+	<!-- 댓글 수정시 UPDATE Query -->
 	<update id="update" parameterType="loungeReplyVO"> 
-		update lounge_reply
-		set
-			reply_content = #{reply_content},
-			reply_date = now()
-		where reply_id = #{reply_id}
+	    UPDATE lounge_reply
+	    SET
+	        reply_content = #{reply_content},
+	        reply_date = NOW()
+	    WHERE reply_id = #{reply_id}
 	</update>
 	
+	<!-- 댓글 삭제시 DELETE Query -->
 	<delete id="delete" parameterType="int">
-		delete from lounge_reply where reply_id = #{reply_id}
+	    DELETE FROM lounge_reply WHERE reply_id = #{reply_id}
 	</delete>
 	
-	<select id="one" parameterType="int" resultType="loungeReplyVO">
-		select * from lounge_reply where reply_id = #{reply_id}
-	</select>
 </mapper>
