+++ src/main/webapp/header.jsp
@@ -125,7 +125,7 @@
                     <div class="nav-item dropdown">
                         <a href="#" class="nav-link dropdown-toggle" data-bs-toggle="dropdown">
                             <img class="rounded-circle me-lg-2" src="/roadpet/resources/img/default.jpg" alt="" style="width: 40px; height: 40px;">
-                            <span class="d-none d-lg-inline-flex"><%=session.getAttribute("user_id")%>님 로그인되었습니다.</span>
+                            <span class="d-none d-lg-inline-flex"><%=session.getAttribute("nickName")%>님 로그인되었습니다.</span>
                         </a>
                        
                         <div class="dropdown-menu dropdown-menu-end bg-light border-0 rounded-0 rounded-bottom m-0">
