--- src/main/webapp/lounge/lounge_index.jsp
@@ -1,65 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-<meta charset="utf-8">
-<title>로드펫 커뮤니티</title>
-<meta content="width=device-width, initial-scale=1.0" name="viewport">
-<meta content="" name="keywords">
-<meta content="" name="description">
-<%@ include file="../sidebar.jsp"%>
-</head>
-<body>
-	<div class="container-fluid position-relative bg-white d-flex p-0">
-		<!-- Spinner Start -->
-		<div id="spinner"
-			class="show bg-white position-fixed translate-middle w-100 vh-100 top-50 start-50 d-flex align-items-center justify-content-center">
-			<div class="spinner-border text-primary"
-				style="width: 3rem; height: 3rem;" role="status">
-				<span class="sr-only">Loading...</span>
-			</div>
-		</div>
-		<!-- Spinner End -->
-
-		<!-- views 파일에서 구현예정 -->
-			<div class="content open">
-				<%@ include file="../header.jsp"%>
-				
-				<h3 >전체글</h3>
-				<div style="display: flex; justify-content: flex-end;">
-				<a href="lounge_insert.jsp">
-					<button type="button"class="btn btn-outline-danger m-2" id="insertType"
-					aria-hidden="true">글쓰기</button> </a>	
-				</div>
-	
-			<!-- Table End -->
-			
-
-		</div>
-		<!-- Content End -->
-
-
-		<!-- Back to Top -->
-		<a href="#" class="btn btn-lg btn-primary btn-lg-square back-to-top"><i
-			class="bi bi-arrow-up"></i></a>
-	</div>
-	<!-- JavaScript Libraries -->
-	<script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
-	<script
-		src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/js/bootstrap.bundle.min.js"></script>
-	<script src="../resources/lib/easing/easing.min.js"></script>
-	<script src="../resources/lib/waypoints/waypoints.min.js"></script>
-	<script src="../resources/lib/owlcarousel/owl.carousel.min.js"></script>
-	<script src="../resources/lib/tempusdominus/js/moment.min.js"></script>
-	<script
-		src="../resources/lib/tempusdominus/js/moment-timezone.min.js"></script>
-	<script
-		src="../resources/lib/tempusdominus/js/tempusdominus-bootstrap-4.min.js"></script>
-
-	<!-- Template Javascript -->
-	<script src="../resources/js/main.js"></script>
-</body>
-
-</html>
